Added basic menu handler and fixed Screen.fill()

+++ core/src/main/java/net/catacombsnatch/game/core/entity/components/Health.java
@@ -21,6 +21,6 @@ public class Health implements EntityComponent {
 	}
 
 	public float getHealthPercentage() {
-		return health / maxHealth;
+		return (float) (health) / (float) (maxHealth);
 	}
 }
